Python projects/self projects/CSV-mysql/csv_to_mysql-main/requirements.txt
PyQt5
mysql-connector-python
pandas
